requirements.txt
Package                Version
---------------------- -------------
apiview                1.3.25
apturl                 0.5.2
asgiref                3.6.0
bcrypt                 3.2.0
blinker                1.4
Brlapi                 0.8.3
certifi                2023.5.7
chardet                4.0.0
charset-normalizer     3.1.0
click                  8.0.3
colorama               0.4.4
command-not-found      0.3
coreapi                2.3.3
coreschema             0.0.4
cryptography           3.4.8
cupshelpers            1.0
dbus-python            1.2.18
defer                  1.0.6
defusedxml             0.7.1
diff-match-patch       20200713
distro                 1.7.0
distro-info            1.1build1
Django                 4.2
django-dotenv          1.4.2
django-filter          23.2
django-import-export   3.2.0
django-rest-auth       0.9.5
djangorestframework    3.14.0
drf-yasg               1.21.5
duplicity              0.8.21
et-xmlfile             1.1.0
fasteners              0.14.1
future                 0.18.2
httplib2               0.20.2
idna                   3.4
importlib-metadata     4.6.4
inflection             0.5.1
itypes                 1.2.0
jeepney                0.7.1
Jinja2                 3.1.2
keyring                23.5.0
language-selector      0.1
launchpadlib           1.10.16
lazr.restfulclient     0.14.4
lazr.uri               1.0.6
lockfile               0.12.2
louis                  3.20.0
macaroonbakery         1.3.1
Mako                   1.1.3
MarkupPy               1.14
MarkupSafe             2.1.2
monotonic              1.6
more-itertools         8.10.0
netifaces              0.11.0
oauthlib               3.2.0
odfpy                  1.4.1
olefile                0.46
openpyxl               3.1.2
packaging              23.1
paramiko               2.9.3
pexpect                4.8.0
Pillow                 9.5.0
pip                    22.3.1
protobuf               3.12.4
psycopg2-binary        2.9.6
ptyprocess             0.7.0
pycairo                1.20.1
pycups                 2.0.1
PyGObject              3.42.1
PyJWT                  2.3.0
pymacaroons            0.13.0
PyNaCl                 1.5.0
pyparsing              2.4.7
pyRFC3339              1.1
python-apt             2.4.0+ubuntu1
python-dateutil        2.8.1
python-debian          0.1.43ubuntu1
pytz                   2023.3
pyxdg                  0.27
PyYAML                 6.0
reportlab              3.6.8
requests               2.30.0
ruamel.yaml            0.17.26
ruamel.yaml.clib       0.2.7
SecretStorage          3.3.1
setuptools             65.5.1
six                    1.16.0
sqlparse               0.4.4
systemd-python         234
tablib                 3.4.0
ubuntu-advantage-tools 8001
ubuntu-drivers-common  0.0.0
ufw                    0.36.1
unattended-upgrades    0.1
uritemplate            4.1.1
urllib3                2.0.2
usb-creator            0.3.7
wadllib                1.3.6
wheel                  0.38.4
xdg                    5
xkit                   0.0.0
xlrd                   2.0.1
xlwt                   1.3.0
zipp                   1.0.0
